+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Go to the `v1` branch to see the changelog of Lume 1.
 
 ## [3.2.5] - Unreleased
 ### Fixed
+- Don't break the build if a loader throws an error.
 - `google_fonts` plugin ignores display parameter [#834]
 - Updated dependencies: `std`, `cliffy`, `pagefind`, `postcss`, `autoprefixer`, `html-validate`, `decap-cms`, `tailwindcss`, `terser` and some icons.